Move refactor into assignment_helper
